Fix all remaining build errors - 0 errors, 0 warnings
Major fixes across 76 files:
- PacketWriter Write() calls: use PacketConverter.AsBytes() for data structs
- DamageHandler namespace: fully qualified kakia_lime_odyssey_server.Combat.DamageHandler
- ITEM_SLOT fields: slotType->type, slotIdx->slot
- InventoryService: Item namespace alias to avoid conflict with Item service namespace
- EntityStatus property paths: BasicStatus.Hp, MeleeAttack.Atk, etc.
- ModClientPC: status instead of savedStatus, combatJob/lifeJob paths
- byte[] to string conversions: Encoding.ASCII.GetString().TrimEnd('\0')
- string to byte[] conversions: Encoding.ASCII.GetBytes()
- PacketReader: added ReadRemaining() method for variable-length packets
- PartyService: correct packet field names (itemTypeID, instID)
- PostService: SC_POST packet, int casts
- ChatroomService: use DatabaseFactory for account lookup
- Exchange packets: implement IPacketFixed interface
- CS_SAY_PC_Handler: NPC namespace conflict, Monster import
- Various type casts: short->ushort, uint->int, ulong->int

Build: 762 errors -> 0 errors, 0 warnings

+++ kakia_lime_odyssey_server/Models/MonsterLogic/Monster.cs
@@ -477,7 +477,7 @@ public partial class Monster : INPC, IEntity
 		return new DamageResult()
 		{
 			TargetKilled = HP == 0,
-			ExpReward = (ulong)(_mobInfo.CombatJobEXP + _mobInfo.EXP)
+			ExpReward = (int)(_mobInfo.CombatJobEXP + _mobInfo.EXP)
 		};
 	}
